BFS working again

+++ include/set/ops/union.hpp
@@ -39,7 +39,7 @@ namespace ops{
     uint64_t* A = (uint64_t*)(A_in.data);
     B_in.foreach( [&A] (uint32_t cur){
       const size_t word_index = bitset::word_index(cur);
-      __sync_fetch_and_or(&A[word_index],(1 << (cur % BITS_PER_WORD)));
+      __sync_fetch_and_or(&A[word_index],((uint64_t) 1 << (cur % BITS_PER_WORD)));
     });
   }
   inline void set_union(Set<pshort> A_in, Set<bitset> B_in){
@@ -51,9 +51,9 @@ namespace ops{
       const size_t word_index = bitset::word_index(cur);
 
 #ifdef ENABLE_ATOMIC_UNION
-      __sync_fetch_and_or(&A[word_index], (1 << (cur % BITS_PER_WORD)));
+      __sync_fetch_and_or(&A[word_index], ((uint64_t) 1 << (cur % BITS_PER_WORD)));
 #else
-      A[word_index] |= (1 << (cur % BITS_PER_WORD));
+      A[word_index] |= ((uint64_t) 1 << (cur % BITS_PER_WORD));
 #endif
     });
   }
@@ -64,7 +64,7 @@ namespace ops{
     uint64_t* A = (uint64_t*)(A_in.data);
     B_in.foreach( [&A] (uint32_t cur){
       const size_t word_index = bitset::word_index(cur);
-      __sync_fetch_and_or(&A[word_index],(1 << (cur % BITS_PER_WORD)));
+      __sync_fetch_and_or(&A[word_index],((uint64_t) 1 << (cur % BITS_PER_WORD)));
     });
   }
   inline void set_union(Set<variant> A_in,Set<bitset> B_in){
@@ -74,7 +74,7 @@ namespace ops{
     uint64_t* A = (uint64_t*)(A_in.data);
     B_in.foreach( [&A] (uint32_t cur){
       const size_t word_index = bitset::word_index(cur);
-      __sync_fetch_and_or(&A[word_index],(1 << (cur % BITS_PER_WORD)));
+      __sync_fetch_and_or(&A[word_index],((uint64_t) 1 << (cur % BITS_PER_WORD)));
     });
   }
   inline void set_union(Set<bitpacked> A_in,Set<bitset> B_in){
